added ToString functionality and resolved errors
